+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.23.2 - 2022-11-29
+
+## Changed
+
+- Add a buffer (of one polling interval) to the start datetime of PR ingestion.
+  This will allow any missed updates to be re-queried and included in the next
+  ingestion round.
+
 ## 1.23.1 - 2022-10-31
 
 ## Changed